Renamed main-store to mainStore as it aligs with js structs
